apps/qwexcli/pyproject.toml
[project]
name = "qwexcli"
dynamic = ["version"]
description = "Qwex CLI - Run orchestration for ML workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["typer>=0.9.0", "rich>=13.0.0", "pydantic>=2.0.0", "pyyaml>=6.0.0"]

[tool.uv.sources]

[project.scripts]
qwex = "qwexcli.main:app"

[tool.uv]
dev-dependencies = [
    "ruff>=0.14.8",
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options.root = "../.."

[tool.hatch.build.hooks.vcs]
version-file = "qwexcli/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["qwexcli"]
